% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPEapp-package.R
\docType{package}
\name{MPEapp-package}
\alias{MPEapp}
\alias{MPEapp-package}
\title{MPEapp: MPE Shiny app (golem)}
\description{
Shiny app built with golem on Zurich rent prices.
}
\author{
\strong{Maintainer}: Sara Hänzi \email{sara.haenzi@zuerich.ch}

Authors:
\itemize{
  \item Sara Hänzi \email{sara.haenzi@zuerich.ch}
}

Other contributors:
\itemize{
  \item Statistik stadt Zürich \email{statistik@zuerich.ch} [copyright holder, funder]
}

}
\keyword{internal}
