% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MPEapp-package.R
\docType{package}
\name{MPEapp-package}
\alias{MPEapp}
\alias{MPEapp-package}
\title{MPEapp: MPE Shiny app (golem)}
\description{
Shiny app built with golem on Zurich rent prices
}
\author{
\strong{Maintainer}: 

Other contributors:
\itemize{
  \item Hänzi Sara \email{sara.haenzi@zuerich.ch}
}

}
\keyword{internal}
